* commit '18b4b60eca2dcfe5f5901ff2552d35766a63e478': (22 commits)
  changed kill browser after command
  removed wip_with_trip
  Remove improper driver name
  removed unnecessary system.out.println and stabilize scenarios
  WIP implementing van inbound, some scenarios are not stable
  Implement all scenarios for shipment van inbound with trip, need check older scenarios
  Implement removed shipment checker, invalid scan scenarios finisehd
  Implement invalid scan transit shpiment from other country
  Adjustment in cancel trip
  Implementing van inbound shipments from another country
  Implement skip some shipment van inbound scenario
  New commit in commons
  Implement shipment van inbound to go with trip scrollable
  Update commons selenium reference, WIP shipment to go with trip
  Update reference for common-selenium
  Implement shipment to go with trip
  New commons reference
  Implement 2 new van inbound with trip scenarios
  Implement scan no path found scenario
  Implement remove shipment scenarios
  ...

+++ src/test/java/co/nvqa/operator_v2/model/ShipmentInfo.java
@@ -29,7 +29,7 @@ public class ShipmentInfo extends DataEntity<ShipmentInfo>
     private String destHubName;
     private String arrivalDatetime;
     private String completedAt;
-    private int ordersCount;
+    private Long ordersCount;
     private String comments;
     private String mawb;
     private String sla;
@@ -50,7 +50,7 @@ public class ShipmentInfo extends DataEntity<ShipmentInfo>
         setDestHubName(shipments.getShipment().getDestHubName());
         setArrivalDatetime(normalisedDate(shipments.getShipment().getArrivalDatetime()));
         setCompletedAt(normalisedDate(shipments.getShipment().getCompletedAt()));
-        setOrdersCount(Math.toIntExact(shipments.getShipment().getOrdersCount()));
+        setOrdersCount((shipments.getShipment().getOrdersCount()));
         setComments(shipments.getShipment().getComments());
         setMawb(shipments.getShipment().getMawb());
     }
@@ -179,12 +179,12 @@ public class ShipmentInfo extends DataEntity<ShipmentInfo>
         this.completedAt = completedAt;
     }
 
-    public int getOrdersCount()
+    public Long getOrdersCount()
     {
         return ordersCount;
     }
 
-    public void setOrdersCount(int ordersCount)
+    public void setOrdersCount(Long ordersCount)
     {
         this.ordersCount = ordersCount;
     }
